+++ app/api/featuredItem/route.tsx
@@ -1,9 +1,9 @@
 import { NextResponse, NextRequest } from "next/server";
-import { RepoApi } from "@/src/utils/apiClients/apiClients";
+import { CollectionsApi } from "@/src/utils/apiClients/apiClients";
 export const fetchCache = "force-no-store";
 
 export const GET = async (request: NextRequest, response: NextResponse) => {
-  const data = await RepoApi.getFeaturedItemData();
+  const data = await CollectionsApi.getFeaturedItemData();
 
   const newResponse = NextResponse.json(data, { status: 200 });